Fall back to searching for filename without ext when updating
When Google Drive imports .csv files it strips the ext from the file
name, so searching for the same name to update the file will end up just
creating a new file and so on. To avoid that, if we don't find full
'name.csv' look for a file called 'name' in parent and update that
before creating a new file.