Better detection of interactive stdin
When run from cron, stdin is not a tty so gsheet would always expect
input making it impossible to read data. This change makes it so gsheet
only expects input when stdin is a tty OR if stdin and stdout are both
pipes and there is no data in stdin.

Should fix #7