Add --read flag and revert v0.1.2 heuristics [BREAKING change]
It turns out that calling FileInfo.Size() on a pipe is NOT a reliable
way to check if there is stdin data available. Because of that the
changed heuristic for detecting input in v0.1.2 broke scripts which rely
on piping data into gsheet.

Because it's probable that more people rely on the original behavior,
this release reverts the new v0.1.2 behavior. It also introduces the
`--read` flag to `csv` which forces gsheet to read data even if stdin is
not a tty. This is the recommended way to read sheets from scripts run
by cron.